refactor: SignupTextFieldTest 분리
